Fixed NPE and failure in monotask tests.
This commit fixes two problems with the tests:
(1) ComputeSchedulerSuite didn't mock the TaskMetrics in the ComputeMonotask
it creates, which led to a NPE in ComputeScheduler when it tries to update the
metrics on task completion. This caused the ComputeSchedulerSuite to fail.

(2) LocalDagSchedulerSuite used a mocked TaskContext and didn't mock the
TaskMetrics, leading to the same issue as above. This didn't cause any test
failures, but led to some annoying extra output with the NPEs when running
the tests.

Both of these problems were introduced by commit 290f86b03, which added
metrics about the time spent queueing (those metrics get updated in
ComputeScheduler, which is the cause of the NPE).