Add support for Tera{Input/Output}Format.
This commit adds support for TeraInputFormat and TeraOutputFormat,
which required adding a case for TeraOutputFormat to
SparkHadoopMapReduceUtil.getRecordWriter(). This alone was not
sufficient, however, because TeraInputFormat caused an error by
constructing multiple FileSplits for each HDFS file, a practice that
Monotasks does not support. In order to prevent TeraInputFormat from
generating too many FileSplits, this commit overrides
FileInputFormat.isSplitable() in TeraInputFormat, hardcoding the method
to always return false.