Throw error if multiple FileSplits per partition.
Currently, when reading HDFS files, Monotasks does not support splitting
an HDFS partition file so that it corresponds to multiple FileSplits
(and therefore multiple Spark partitions). There must be a one-to-one
mapping between HDFS partitions and Spark partitions. If this invariant
does not hold, then Monotasks will silently drop records. This is not a
fundamental property of the Monotasks design, we just have not
implemented support for multiple splits per HDFS partition yet.

This commit modifies NewHadoopRDD.getPartitions() in order to verify
that each split corresponds to an entire HDFS file partition.