Added support for reading Parquet files from HDFS.
This commit modifies NewHadoopRDD so that it is capable of reading HDFS
files stored in the Parquet columnar storage format. This required
adding a special case in NewHadoopRDD to handle ParquetInputSplits
differently than generic FileSplits. Furthermore, this commit adds the
ParquetInputSplitPrivateMethodAccessor class, which provides access to
package-private methods in the ParquetInputSplit class. In order to use
the Parquet libraries in spark-core, a dependency on
com.twitter.parquet-hadoop was added to core/pom.xml.