Don't tell master about data cached by DiskReadMonotask.
This commit changes DiskReadMonotask so that it sets tellMaster to
false when storing its result in memory. This is because DiskReadMonotask
is always used to read intermediate data into memory; that data will be
dropped from memory as soon as the DiskReadMonotask's dependents complete, so
the master shouldn't be told about it.