Add variation to compute times
This commit introduces the "compute_variation" configuration parameter,
which adds a degree of randomness to the duration of a ComputeMonotask.
This change is required to simulate how variations in the runtime of
supposedly identical tasks lead to periods of time where the CPU is
idle, even though it is the bottleneck resource.