Improved monotasks scheduling interface.
This commit adds a new class, MonotasksScheduler, that can be used to implement
different policies for how monotasks are assigned to worker machines.
MonotasksSchedulers are responsible for making an initial allocation of a stage's
tasks to worker machines, and then on the worker, are responsible for requesting
additional tasks from the master. Unlike the previous scheduling mechansim, all
new tasks to run on the worker after the initial allocation will be requested by
the worker machine (so for example, a task finishing on a worker no longer triggers
a new task to be launched by default).