Add script to simulate in-memory shuffle job
This commit adds a "simulated_workloads" directory and a script that
simulates an in-memory shuffle job. The cluster parameters are hardcoded
in the script, and only job parameters like the ratio of network time to
compute time are available as command line arguments. The purpose of
this script is to allow quick testing of different job parameters
without needing to create a new XML configuration file for each trial.