Write simulation info to a file
The commit modifies the Simulator to save the simulation configuration,
macrotask distribution information, and job completion time information
to a file once a simulation has finished. This file is colocated with
the continuous monitor log files. The purpose of this change is to
record for easy reference the exact parameters that produced a set of
logs.