Allow PrepareMonotasks to jump the queue.
With this commit, prepare monotasks are always placed at the front
of the queue of CPU monotasks. The reason for doing this is that
mid-way through a CPU-contended job, we want to make sure the
(very short) prepare monotasks can run as soon as a task is assigned
to a machine, so the macrotasks associated with those prepare monotasks
can pipeline use of other resources with the CPU. Otherwise, if the
prepare monotasks are placed at the end of the queue, the CPU can
end up idle (even in a CPU-bottlenecked job) between when the prepare
monotasks complete and when some other (non-bottleneck) resource, like
disk or network, is used.