Updated SortJob to force a GC between jobs.
This behavior is consistent with the behavior in MemorySortJob, and helps to
avoid garbage collection while tasks are running.

This commit also fixes both jobs to use a number of GC tasks exactly
equal to the number of executors (getExecutorStorageStatus returns one
more status than the number of executors, because there's one status for
the master).