Remove default AMI from spark_ec2.py
Currently, the spark_ec2.py script always uses the same AMI unless the
user explicitly specifies a different one. This default AMI is a PVM
AMI, and is therefore not compatible with instance types that require an
HVM AMI. To resolve this incompatibility, this commit removes the
hardcoded default AMI value, causing the spark_ec2.py script to use
Spark's default AMI for a particular instance type when the user does
not specify an AMI manually. The main side effect of this change is that
the previously-hardcoded AMI had a 16GB root volume, whereas Spark's
default AMIs only have an 8GB root volume.