Add num threads per disk experiment
This commit adds a script that runs a series of experiments that write
and read a large amount of data using different values of the
configuration parameter "spark.monotasks.threadsPerDisk". The goal of
this script is to help us determine the optimal value of
"spark.monotasks.threadsPerDisk" for clusters that use SSDs instead of
HDDs.