Adapt task scheduler for multiple threads per disk
This commit is a hack that modifies a worker machine's task slot
calculation to account for the fact that now there may be many threads
accessing each disk. The ability to have multiple threads per disk means
that the effective disk concurrency is the number of disks times and
number of threads per disk.