Fix build_ssh_command() username for EC2
Previously, the experiment_scripts/utils.build_ssh_command() function
did not include a username in the SSH command that it produced. This
does not work in the case where the command is intended to log in to an
EC2 machine from a local machine. Therefore, this commit changes
build_ssh_command() to use the username "root" when logging in to EC2
machines.