Refactor BDB experiment script to run on master
Previously, the run_bdb_experiments.py script was designed to be run
from a machine other than the master node of the cluster that it was
exercising. All commands were executed via SSH, meaning that the machine
running the script required a reliable network connection for the
entirety of the (sometimes lengthy) experiments. This proved to be very
inconvenient. Therefore, this commit refactors the
run_bdb_experiments.py script so that it is now meant to be run on the
master node directly. In addition, instead of only collecting the
continuous monitor from one worker, the script now collects all
continuous monitors.