Added start and end disk and network counters.
This commit adds the start and end disk and network counters
while tasks were running to the TaskMetrics, and to the JSON
event log output. This is useful for computing the I/O
throughput while an entire job was running, and for determining
how much time disk and network monotasks took (we can't directly
use the monotask time, because that time doesn't account for the
fact that many monotasks may have been running concurrently).