Convert all shuffled pairs to tuples.
This commit adds code in ShuffleMapMonotask to ensure that all shuffle
data is written as a Tuple2, rather than as a Product2 (which is a
parent class of Tuple2). This is to fix an inconsitency with Spark,
where Spark converts everything to Tuple2s, which means it uses a
different serialization path.