Add intrumentation for network transmit idle time
This commit adds a counter to the NetworkScheduler that tracks the total
nanoseconds during which no data is being transmitted by
NetworkResponseMonotasks. This required also adding a set that keeps
track of the currently running NetworkResponseMonotasks. The purpose of
this new counter is to help us determine the actual network throughput
of the worker. This counter is included in the TaskMetrics, where is it
set when a task completes, so that it is logged in the Event Log.