Remove sorted data when SortJob starts.
The current SortJob doesn't delete any of the sorted data it
generates, which can cause the disk to fill up.  This commit
changes that job so it deletes any old sorted data before a new
experiment starts (this both helps with disks-filling-up issues,
and also avoids scenarios where the job fails because there's
already a file with the filename that the job is using for the
output data).