Implemented priority queue for disk monotasks.
With this change, disk monotasks are run in order based on the
ID of the associated macrotask. Macrotask IDs are assigned in the
order that the scheduler assigns tasks to machines, so this change
effectively implements SRPT for monotasks. This change is beneficial
because it helps to ensure that a machine's local tasks cannot
accumulate in the disk queue in a way that causes requests from other
machines to be starved.