Adjust log level of log messages.
This commit demotes a bunch of log messages from INFO to DEBUG
level because the size of the info-level logs was getting to be
very large.  We probably need to do some more demotion (especially
of some messages in the BlockManager relating to shuffle blocks)
but this is a start.

This commit also changes one message that's about tasks failing
to ERROR level.