Separate HDFS serialization and compression
This commit breaks the pipelining of HDFS data deserialization/
decompression and serialization/compression. If the
"spark.monotasks.separateHdfsSerialization" configuration parameter is
set to "true", then the deserialization/decompression and serialization/
compression of HDFS data will be separated from the process of computing
an RDD so that its runtime can be measured separately. This commit also
adds fields to TaskMetrics to track the time spent in deserialization/
decompression and serialization/compression.