Fix HDFS serialization separation
This commit fixes the HDFS (de)serialization/(de)compression separation
code to account for the fact that HDFS reuses the same Writable object
for every element in a dataset. This means that in order to break the
pipelining, we need to copy each value as we step through the iterator.
Additionally, this commit modifies the InsertIntoHiveTable class to
support separating (de)serialization/(de)compression when creating Hive
tables.