Throw an error if task reads remote data.
Monotasks doesn't currently support reading remote RDD partitions (except
in the case of shuffle data). Previously, this meant that if a monotask
needed data that was only stored remotely, it would re-compute the data.
This is not the desired behavior and leads to hard-to-interpret experimental
results; to avoid this, this change throws an error when a task tries to read
data that is only stored remotely.