Fix bug in determining tasks' resource needs.
The existing code only looked at the final RDD a particular stage
computed to determine tasks' resource needs. This commit fixes the
code to walk back through all of the narrow dependencies to determine
if any RDDs that will be computed as part of the stage require reading
from or writing to disk.