Add option to opportunistically begin shuffle early.
This commit adds a parameter spark.monotasks.earlyShuffle that, if true,
means that worker machines will begin sending the shuffle data over the
network before map tasks have completed.

There are a few changes to enable the new functionality:
-When the scheduler creates a map stage, it also generates a tentative
mapping of reduce tasks to executors, and passes this to the map tasks.
The map tasks use this to determine where to send shuffle data. If this
mapping ends up needing to change when the reduce tasks are actually
scheduled, we're no worse off than if no data had been sent opportunistically,
because the reduce task still needs to fetch all data remotely.
-There's now a low priority queue in the NetworkScheduler that's used to
opportunistically begin sending shuffle data. Network monotasks are only
run from the low priority queue when there are no high priority tasks to run.
One complexity of doing this is that there are some race conditions where
a concurrently running low and high priority monotask may be fetching
the same shuffle block. To handle this, the TransportResponseHandler checks, when
a new block is requested, that there's not already an outstanding request for it.
If there is an outstanding request, it fails the low priority request, so that
the BlockReceivedCallback for the high priority monotask (which is the one
that a task depends on) is the one that gets executed when the block arrives.

One more minor change is that the ShuffleWriter interface was changed to
return the size of each output block (rather than a map status, which has
compressed sizes), so that these sizes may be sent to reduce tasks.

One issue with this change is that it can result in twice the amount of
shuffle data being stored in-memory, because the shuffle data that's
opportunistically fetched isn't deleted until the reduce stage ends.