Refactor the way that local and remote macrotasks are tracked.
This commit alters how local and remote macrotasks are tracked: there's
now an explicit set of local macrotasks, rather than tracking a set of
all macrotasks, and then subtracting the number of remote macrotasks to
determine the number of local monotasks. This change is in anticipation
of a future change that alters the way that shuffle data is fetched;
with that change, the existing tracking didn't work, because a single
macrotask could be double counted (once as a local macrotask and once
as a remote macrotask).