Avoid unnecessary RPC in PrepareMonotasks.
Calling BlockManager::getAllStorageLocations causes an RPC to the master,
which can be slow, especially when the master is busy launching other
tasks (in which case the RPC seems to get stuck in a long queue). This
commit temporarily disables that call (which is only used to check for
an error check) to avoid the added latency.