Prioritize result serialization monotasks.
Like the PrepareMonotasks, these monotasks block important work from happening,
so should be prioritized (otherwise, for example, they can keep new monotasks
from getting assigned to the machine, which leads to CPU utilization problems
later on).

This is kind of a hack, and should likely be fixed later as part of a more
holistic ordering of monotasks.