Order low-priority network monotasks by reduce ID.
This commit changes the NetworkScheduler to run low-priority network monotasks in order
of reduce ID, so that the data for the first reduce task is all fetched before data
for later reduce tasks (and so-on). This helps to avoid the ramp-up time at the
beginning of the reduce phase, where the first task gets stuck still waiting for
some of its data.