Cap outstanding network monotasks by bytes.
Previously, the number of outstanding network monotasks was limited by
a maximum number of macrotasks that could have outstanding tasks at a
time. This commit changes that so that instead, the number of oustanding
network monotasks is limited by a target amount of outstanding bytes. This
simplifies the code, is more robust to different task sizes, and fixes
some prioritization issues with low priority monotasks.