Loosen locality constraints for reduce tasks.
A previous commit assigned reduce tasks to machines before map tasks
run, so map tasks can begin sending output to reduce tasks before they
begin. These assignments were used as locality constraints when assigning
the reduce tasks to machines, to ensure that reduce tasks were placed on
the same machines where data was being sent.

That assignment of tasks to machines was not always ideal: we evenly
balance reduce tasks across machines, but often machines have different
speeds, so the fastest machines can process more than their share of reduce
tasks.  This commit changes the scheduler to treat the reduce task locality
preferences as suggestsions but not requirements: the scheduler tries to
satisfy them, but will schedule tasks elsewhere when the locality constraints
can't be satisfied.

This commit does not cancel fetch requests that may be in progress to fetch
data for a reduce task on a particular executor that's different where the
reduce task ended up running.